requirements.txt
asgiref==3.7.2
async-generator==1.10
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
boto3==1.26.160
botocore==1.29.160
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chompjs==1.2.2
constantly==15.1.0
crispy-bootstrap4==2022.1
cryptography==41.0.1
cssselect==1.2.0
Django==4.2.2
django-cors-headers==4.1.0
django-crispy-forms==2.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
filelock==3.12.2
GDAL==3.2.2.1
gunicorn==20.1.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.2
outcome==1.2.0
packaging==23.1
parsel==1.8.1
Protego==0.2.1
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
PyJWT==2.7.0
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-scrapyd-api==2.1.2
pytz==2023.3
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
s3transfer==0.6.1
Scrapy==2.9.0
scrapy-djangoitem==1.1.1
scrapyd==1.4.2
selenium==4.10.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tldextract==3.4.4
trio==0.22.0
trio-websocket==0.10.3
Twisted==22.10.0
typing_extensions==4.6.3
urllib3==1.26.16
w3lib==2.1.1
wsproto==1.2.0
zope.interface==6.0
